+++ MSETx.h
@@ -132,7 +132,7 @@ fun ok64 X(MSET, EjectAtZ)(X(, css) heap, size_t at, X(, z) z) {
         h[last][1] = t1;
     }
     --heap[1];
-    if (at < $len(heap)) {
+    if ((i64)at < $len(heap)) {
         X(MSET, _UpZ)(heap, at, z);
         X(MSET, _DownZ)(heap, at, z);
     }
@@ -242,12 +242,12 @@ fun ok64 X(MSET, Compact)(X(, css) stack, X(, s) into) {
     if (n < 2) return OK;
     size_t m = 1;
     size_t total = $len(stack[0][n - 1]);
-    while (m < n && total * 8 > $len(stack[0][n - 1 - m])) {
+    while (m < n && (i64)(total * 8) > $len(stack[0][n - 1 - m])) {
         total += $len(stack[0][n - 1 - m]);
         m++;
     }
     if (m < 2) return OK;
-    if ($len(into) < total) return MSETNOROOM;
+    if ($len(into) < (i64)total) return MSETNOROOM;
     T *base = *into;
     X(, css) sub = {stack[0] + (n - m), stack[0] + n};
     X(MSET, Start)(sub);
